Rename `validate.py` -> `core.py` bc neatness
