+++ pydian/validation/__init__.py
@@ -1,13 +1,13 @@
+from .core import validate
 from .rules import RC, RGC, Rule, RuleGroup
 from .specific import InRange, InSet, IsOptional, IsRequired, IsType, MaxCount, MinCount
-from .validate import validate
 
 __all__ = [
     "RC",
     "RGC",
     "Rule",
     "RuleGroup",
-    "validate",
+    "core",
     "IsOptional",
     "IsRequired",
     "InRange",
