+++ tests/validation/test_core_validation.py
